requirements.txt
django
djangorestframework
psycopg2
django-debug-toolbar
django-cors-headers
